+++ backend/src/main/java/com/attiatlttofafrn/backend/repository/TaskRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.attiatlttofafrn.backend.model.Task;
 import com.attiatlttofafrn.backend.model.User;
 
-public interface TaskRepository extends JpaRepository<Task, Integer> {
+public interface TaskRepository extends JpaRepository<Task, Long> {
 
     List<Task> findByUser(User user);
 }
