+++ backend/src/main/java/com/attiatlttofafrn/backend/repository/UserRepository.java
@@ -7,7 +7,8 @@ import org.springframework.stereotype.Repository;
 
 import com.attiatlttofafrn.backend.model.User;
 
-@Repository 
-public interface  UserRepository extends  JpaRepository<User, Integer>{
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
     Optional<User> findByEmail(String email);
 }
